Revert "Refactor FormikFileInput to always upload using storage context."
This reverts commit 938e958b776ca4773c15a72cae52830e2b89b83d.
