+++ src/core/ui/upload/FileUpload/FileUpload.tsx
@@ -5,7 +5,7 @@ import { useTranslation } from 'react-i18next';
 import 'react-image-crop/dist/ReactCrop.css';
 import UploadButton from '../../button/UploadButton';
 import { StorageConfig } from '../../../../domain/storage/StorageBucket/useStorageConfig';
-import { useUploadFileMutation } from '../../../apollo/generated/apollo-hooks';
+import { useUploadFileMutation, useUploadFileOnReferenceMutation } from '../../../apollo/generated/apollo-hooks';
 import { useNotification } from '../../notifications/useNotification';
 
 interface FileUploadProps {
@@ -16,12 +16,19 @@ interface FileUploadProps {
 
 const bytesInMegabyte = Math.pow(1024, 2);
 
-const FileUploadButton: FC<FileUploadProps> = ({ onUpload, storageConfig }) => {
+const FileUploadButton: FC<FileUploadProps> = ({ onUpload, referenceID, storageConfig }) => {
   const { t } = useTranslation();
   const notify = useNotification();
 
   const acceptedFileTypes = storageConfig.allowedMimeTypes.join(',');
   const maxFileSizeMb = storageConfig.maxFileSize ? storageConfig.maxFileSize / bytesInMegabyte : 0;
+
+  const [uploadFileOnReference, { loading: loadingOnReference }] = useUploadFileOnReferenceMutation({
+    onCompleted: data => {
+      notify(t('components.file-upload.file-upload-success'), 'success');
+      onUpload?.(data.uploadFileOnReference.uri);
+    },
+  });
 
   const [uploadFile, { loading: loadingOnStorageBucket }] = useUploadFileMutation({
     onCompleted: data => {
@@ -29,7 +36,7 @@ const FileUploadButton: FC<FileUploadProps> = ({ onUpload, storageConfig }) => {
       onUpload?.(data.uploadFileOnStorageBucket);
     },
   });
-  const loading = loadingOnStorageBucket;
+  const loading = loadingOnReference || loadingOnStorageBucket;
 
   const handleSubmit = async (selectedFile: File) => {
     if (!selectedFile) return;
@@ -38,14 +45,25 @@ const FileUploadButton: FC<FileUploadProps> = ({ onUpload, storageConfig }) => {
       notify(t('components.file-upload.file-size-error', { limit: maxFileSizeMb }), 'error');
       return;
     }
-    await uploadFile({
-      variables: {
-        file: selectedFile,
-        uploadData: {
-          storageBucketId: storageConfig.storageBucketId,
+    if (referenceID) {
+      await uploadFileOnReference({
+        variables: {
+          file: selectedFile,
+          uploadData: {
+            referenceID,
+          },
         },
-      },
-    });
+      });
+    } else {
+      await uploadFile({
+        variables: {
+          file: selectedFile,
+          uploadData: {
+            storageBucketId: storageConfig.storageBucketId,
+          },
+        },
+      });
+    }
   };
 
   return (
